+++ CHANGELOG.md
@@ -92,6 +92,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Results URL in project
 - Results already sent button
 - Votum in project view overview with only one bridgehead
+- Votum for all
 
 ### Changed
 - Rename accept and reject bridgehead state buttons (authorize/revoke)